fix(hex): widen boundary rule coverage and close cross-module infra leaks
The R1/R2/R4/R5 globs used a single `*`, so they never matched the
depth-2 hexagonal sub-contexts under modules/accounting/ (financial-
statements, trial-balance, worksheet, initial-balance, equity-statement).
~24 files carried 31 real violations that the linter could not see.

Widening to `modules/**/` surfaces them: 107 -> 138 no-restricted-imports
on the same source tree. The count going UP is the point — the previous
number was not a smaller problem, it was a blinder.

- Widen the three hexagonal `files` globs to `modules/**/`
- Exclude `modules/**/__tests__/{domain,application,presentation}/**`,
  fixtures shaped like layers. The segment order matters: a naive
  `__tests__/**` also matches `domain/__tests__/` and silently drops 22
  real violations while the total reads like an improvement.
- Deepen the composition-root exemption so the 5 nested roots are not
  flagged as false positives

Also give modules/ai-agent the composition-root.ts that the other 32
modules already have, per docs/architecture/03-rules-hard-rules.md:31.
Its wiring lived in server.ts, the public barrel, which is what R4 exists
to protect. 9 -> 2 R4 violations there.

And close the cross-module encapsulation break where ai-agent and
documents reached into modules/tags/infrastructure directly, following
the operational-doc-type precedent (re-export through the owning
module's barrel).

Both remaining ai-agent violations are the analyzeDocument leak, left
visible on purpose rather than moved into the exempt composition-root.

tsc clean. Tests: 7323 pass, 3 pre-existing failures in
payment-form-shortcut.test.tsx.
