+++ modules/ai-agent/presentation/server.ts
@@ -8,9 +8,10 @@ import "server-only";
  * client-marked module transitively imports this file. Codebase precedent:
  * features/ai-agent/server.ts L1 (pre-migration).
  *
- * Composition root: zero-arg factory `makeAgentService()` per design §3 (D1).
- * Paired sister: `modules/dispatch/presentation/composition-root.ts`
- * (dispatch C3 GREEN `5fd0cd42`) — same shape, single point of adapter wiring.
+ * Wiring lives in `./composition-root` (the ONE file allowed to import
+ * `infrastructure/` per R4, docs/architecture/03-rules-hard-rules.md:31).
+ * This barrel only re-exports — same convention as the other 32 modules.
+ * Paired sister: `modules/dispatch/presentation/server.ts`.
  *
  * AXIS-DISTINCT (honest surface): this is the FIRST hex module with a
  * dual-barrel presentation (server.ts + client.ts) — D5 lock per design §4.
@@ -20,77 +21,8 @@ import "server-only";
  * so consumer cutover at C4 is a pure path rewrite (no surface drift).
  */
 
-import { CerebrasLLMAdapter } from "../infrastructure/llm/cerebras-llm.adapter";
-import { PrismaChatMemoryRepository } from "../infrastructure/prisma/prisma-chat-memory.repo";
-import { PrismaAgentContextRepository } from "../infrastructure/prisma/prisma-agent-context.repo";
-import { PrismaRateLimitRepository } from "../infrastructure/prisma/prisma-agent-rate-limit.repo";
-import { LegacyAccountsAdapter } from "../infrastructure/legacy-accounts.adapter";
-import { LegacyRagAdapter } from "../infrastructure/legacy-rag.adapter";
-import { PrismaTagsRepository } from "@/modules/tags/infrastructure/prisma/prisma-tags.repository";
-import { makeRagService } from "@/modules/rag/presentation/server";
-import {
-  LocalLotInquiryAdapter,
-  makeLotService,
-} from "@/modules/lot/presentation/server";
-import { AgentService } from "../application/agent.service";
-import { AgentRateLimitService } from "../application/rate-limit.service";
-import { PricingService } from "../application/pricing/pricing.service";
-
-/**
- * Composition root — zero-arg factory per design §3 / D1.
- * Mirror of paired sister `makeDispatchService()` (dispatch C3 GREEN `5fd0cd42`).
- *
- * Runtime invocation requires DB + CEREBRAS_API_KEY env — at import time
- * only the factory function shape is exposed (no top-level side effects
- * outside the CerebrasLLMAdapter module load, which validates
- * CEREBRAS_API_KEY at module init mirror of pre-cutover Gemini behavior).
- * GeminiLLMAdapter is preserved in tree (multi-provider fallback) but
- * unused at the composition root.
- *
- * Cleanup #2026-05-17 — `AccountingQueryAdapter` y los 5 services contables
- * (journals/ledger/sale/purchase/payments/accounts/contacts) que solo se
- * usaban acá fueron retirados. Las 8 tools de consulta contable del
- * sidebar-qa duplicaban las páginas dedicadas; Marco decidió que la UI
- * exacta es más confiable que la respuesta del LLM. Los services siguen
- * siendo invocados desde sus propios módulos / otros consumers.
- */
-export function makeAgentService(): AgentService {
-  const llmProvider = new CerebrasLLMAdapter();
-  const chatMemory = new PrismaChatMemoryRepository();
-  const contextReader = new PrismaAgentContextRepository();
-  const rateLimitRepo = new PrismaRateLimitRepository();
-  const accountsLookup = new LegacyAccountsAdapter();
-  // REQ-43 — LegacyRagAdapter accepts an optional TagsRepositoryPort so
-  // searchDocuments tag filters can resolve slugs to tag IDs server-side.
-  // F4 — LegacyRagAdapter takes RagService as a REQUIRED param; this
-  // composition root is where modules/rag gets wired into the agent.
-  const rag = new LegacyRagAdapter(makeRagService(), new PrismaTagsRepository());
-  const lotInquiry = new LocalLotInquiryAdapter(makeLotService());
-  const pricingService = new PricingService();
-  const rateLimit = new AgentRateLimitService(rateLimitRepo);
-
-  return new AgentService({
-    llmProvider,
-    chatMemory,
-    contextReader,
-    rateLimit,
-    accountsLookup,
-    rag,
-    lotInquiry,
-    pricingService,
-  });
-}
-
-/**
- * Composition root — zero-arg factory for AgentRateLimitService.
- * Wires PrismaRateLimitRepository → AgentRateLimitService. Mirror of
- * makeAgentService shape — preserves zero-arg call site at consumers
- * after C4 cutover (old `new AgentRateLimitService()` was zero-arg).
- */
-export function makeAgentRateLimitService(): AgentRateLimitService {
-  const rateLimitRepo = new PrismaRateLimitRepository();
-  return new AgentRateLimitService(rateLimitRepo);
-}
+// ── Composition-root factories ──────────────────────────────────────────────
+export { makeAgentService, makeAgentRateLimitService } from "./composition-root";
 
 // ── Service re-exports (orchestrators) ──────────────────────────────────────
 export { AgentService } from "../application/agent.service";
@@ -116,20 +48,26 @@ import {
 import type {
   FindAccountsResult,
 } from "../domain/tools/tool-output.types";
+import { makeAccountsLookup } from "./composition-root";
 
 /**
- * Composition-root wrapper for executeFindAccountsByPurpose — preserves the
- * pre-migration two-arg call site at `app/api/.../accounts/route.ts`
- * (consumer calls `executeFindAccountsByPurpose(orgId, args)` with no deps).
- * Wires LegacyAccountsAdapter at the composition boundary so the consumer
- * stays barrel-clean per D1 zero-arg factory convention.
+ * Wrapper for executeFindAccountsByPurpose — preserves the pre-migration
+ * two-arg call site at `app/api/.../accounts/route.ts` (consumer calls
+ * `executeFindAccountsByPurpose(orgId, args)` with no deps).
+ *
+ * This is NOT pure wiring: the optional `deps` override lets consumers/tests
+ * inject fakes, so it stays in the barrel. The "which concrete adapter is
+ * the default" decision is delegated to `makeAccountsLookup()` in the
+ * composition root, keeping infrastructure out of this file (R4).
+ * Construction stays lazy — identical runtime behavior to the previous
+ * `deps?.accountsLookup ?? new LegacyAccountsAdapter()`.
  */
 export async function executeFindAccountsByPurpose(
   organizationId: string,
   input: { purpose: "expense" | "bank" | "cash"; query?: string },
   deps?: Partial<FindAccountsByPurposeDeps>,
 ): Promise<FindAccountsResult> {
-  const accountsLookup = deps?.accountsLookup ?? new LegacyAccountsAdapter();
+  const accountsLookup = deps?.accountsLookup ?? makeAccountsLookup();
   return _executeFindAccountsByPurpose(organizationId, input, {
     accountsLookup,
     orgSettingsService: deps?.orgSettingsService,
@@ -153,4 +91,7 @@ export type {
 // Bypasses LLMProviderPort by design — Gemini's analyzeDocument SDK call
 // shape differs from LLMClient.query; 1 consumer (app/api/analyze/route.ts).
 // Documented as honest residual in archive (paired sister: dispatch HubService).
+// Deliberately NOT laundered through composition-root: this is not wiring,
+// it is presentation re-exporting an infrastructure function. It stays here,
+// visible to ESLint (R4), until a proper port/use-case is introduced.
 export { analyzeDocument } from "../infrastructure/llm/gemini-llm.adapter";
